requirements.txt
asgiref==3.2.10
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.45.0
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.12.0
django-filter==21.1
django-tables2==2.4.0
filelock==3.0.12
fire==0.1.2
flake8==3.8.4
geographiclib==1.52
geopy==2.2.0
idna==3.2
mccabe==0.6.1
numpy==1.21.3
oauthlib==3.1.1
peewee==2.8.8
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.1.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.3.1
urllib3==1.26.6
virtualenv==20.8.0
whitenoise==5.3.0
Workon==0.8.2
reportlab
python-dateutil
pdfkit
